chatbot-csv-qa/requirements.txt
flask==2.3.2
pandas==2.2.2
numpy==1.26.0
sentence-transformers==3.0.1
transformers==4.41.2
torch==2.8.0
faiss-cpu==1.11.0
python-dotenv==1.0.0
faker==17.3.0
pillow==9.5.0